update to latest fitbit_api
This also adds the `multi_json` gem to the Gemfile.

The `omniauth-fitbit` gem assumes that the multi_json gem is available
without explicitly defining it as a dependency [1]. As of fitbit_api 1.0
the oauth2 dependency has been bumped to 2.0 which no longer pulls in
the multi_json gem.

This includes the multi_json gem so that the omniauth-fitbit gem can
function properly. That gem should ideally be updated to either set
multi_json as a dependency of it or remove its use entirely, however,
given that it has not been updated in years this change seems unlikely.

[1] https://github.com/tkgospodinov/omniauth-fitbit/blob/acb6408ec2a3713562ed5902f0993b73e2f70125/lib/omniauth/strategies/fitbit.rb#L65